shiny-bakehouse/requirements.txt
shiny
databricks-sql-connector
pandas
plotly
shinywidgets
posit-sdk